litestarApp/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asynch==0.2.5
black==24.10.0
certifi==2024.12.14
charset-normalizer==3.4.1
ciso8601==2.3.2
click==8.1.8
clickhouse-driver==0.2.9
clickhouse-sqlalchemy==0.3.2
Faker==35.0.0
fastapi==0.115.6
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
leb128==1.0.8
litestar==2.14.0
litestar-htmx==0.4.1
lz4==4.3.3
markdown-it-py==3.0.0
mdurl==0.1.2
msgspec==0.19.0
multidict==6.1.0
multipart==1.2.1
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
polyfactory==2.18.1
psycopg2-binary==2.9.10
pydantic==2.10.5
pydantic_core==2.27.2
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
rich-click==1.8.5
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.37
starlette==0.41.3
typing_extensions==4.12.2
tzlocal==5.2
urllib3==2.3.0
uvicorn==0.34.0
zstd==1.5.6.2
